+++ LazZiya.ImageResize/ImageWatermarkOptions.cs
@@ -1,10 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Drawing;
-using System.Text;
+﻿using System.ComponentModel.DataAnnotations;
 
-namespace LazZiya.ImageResize.Watermark
+namespace LazZiya.ImageResize
 {
     /// <summary>
     /// Define options for adding text watermark over the image, like text color, opacity, text outline, etc.
