+++ LazZiya.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 16.0.29509.3
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LazZiya.ImageResize", "LazZiya.ImageResize\LazZiya.ImageResize.csproj", "{23C49B9A-D554-4F5B-8201-5F5C69FF41C0}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebApplication1", "..\Test\WebApplication1\WebApplication1.csproj", "{2E102896-7656-41A6-B4AC-97B09D6E9F5E}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -23,14 +21,6 @@ Global
 		{23C49B9A-D554-4F5B-8201-5F5C69FF41C0}.Release|Any CPU.Build.0 = Release|Any CPU
 		{23C49B9A-D554-4F5B-8201-5F5C69FF41C0}.Release|x64.ActiveCfg = Release|x64
 		{23C49B9A-D554-4F5B-8201-5F5C69FF41C0}.Release|x64.Build.0 = Release|x64
-		{2E102896-7656-41A6-B4AC-97B09D6E9F5E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2E102896-7656-41A6-B4AC-97B09D6E9F5E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2E102896-7656-41A6-B4AC-97B09D6E9F5E}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{2E102896-7656-41A6-B4AC-97B09D6E9F5E}.Debug|x64.Build.0 = Debug|Any CPU
-		{2E102896-7656-41A6-B4AC-97B09D6E9F5E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2E102896-7656-41A6-B4AC-97B09D6E9F5E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2E102896-7656-41A6-B4AC-97B09D6E9F5E}.Release|x64.ActiveCfg = Release|Any CPU
-		{2E102896-7656-41A6-B4AC-97B09D6E9F5E}.Release|x64.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
